refactor: 모델 Gemini-2.5-Flash로 변경

+++ packages/backend/infrastructure/actions/syncPerformanceData.ts
@@ -74,15 +74,25 @@ export const syncPerformanceData = async (
   const { idsToDelete, idsToInsert } = await compareNewOld(newPerformances);
 
   // 공연 데이터 삭제
-  console.log(`IDs to Delete: ${idsToDelete}`);
-  logger.info("Deleting old performance datas...");
+  if (idsToDelete.length > 0) {
+    console.log(`🚀 IDs to Delete: ${idsToDelete.length}`);
+    logger.info("Deleting old performance datas...");
 
-  // 내부에서 fallback 로직 실행
-  await deletePerformances(idsToDelete);
+    // 내부에서 fallback 로직 실행
+    await deletePerformances(idsToDelete);
+  } else {
+    logger.info("Nothing to Delete.");
+  }
 
-  logger.info("Processing New & Updated Performance datas...");
+  if (idsToInsert.length > 0) {
+    console.log(`🚀 IDs to Insert: ${idsToInsert.length}`);
+  } else {
+    logger.info("Noting to Insert.");
+  }
 
   // 수정된 공연 데이터 가져오기
+  logger.info("Processing New & Updated Performance datas...");
+
   const idsToUpdate = await getPerformanceIds(
     startDate,
     updateEndDate,
